[enhancement] More work on the SplashScreen #21
