+++ pom.xml
@@ -19,7 +19,7 @@
             **/waterworld/**/*,
             **/yaeger/guice/**/*,
             **/yaeger/javafx/**/*,
-            **/yaeger/screens/**/*
+            **/yaeger/engine/scenes/splash/**/*,
         </sonar.exclusions>
     </properties>
 
@@ -98,12 +98,10 @@
                         <exclude>nl/meron/showcase/**/*</exclude>
                         <exclude>nl/meron/waterworld/**/*</exclude>
                         <exclude>nl/meron/yaeger/guice/**/*</exclude>
-                        <exclude>nl/meron/screens/**/*</exclude>
-                        <exclude>nl/meron/screens/entities/**/*</exclude>
-                        <exclude>nl/meron/screens/timers/**/*</exclude>
                         <exclude>nl/meron/yaeger/javafx/debug/*</exclude>
                         <exclude>nl/meron/yaeger/javafx/image/*</exclude>
                         <exclude>nl/meron/yaeger/javafx/animationtimer/*</exclude>
+                        <exclude>nl/meron/yaeger/engine/scenes/splash/**/*</exclude>
                     </excludes>
                 </configuration>
             </plugin>
@@ -130,7 +128,6 @@
                 <version>3.2.0</version>
                 <configuration>
                     <excludePackageNames>nl.meron.pong</excludePackageNames>
-                    <excludePackageNames>nl.meron.screens</excludePackageNames>
                     <excludePackageNames>nl.meron.showcase</excludePackageNames>
                     <excludePackageNames>nl.meron.waterworld</excludePackageNames>
                     <source>${java.version}</source>
