+++ src/main/java/nl/meron/yaeger/engine/scenes/splash/entities/Logo.java
@@ -1,11 +1,11 @@
-package nl.meron.yaeger.screens.splash.entities;
+package nl.meron.yaeger.engine.scenes.splash.entities;
 
 import nl.meron.yaeger.engine.Size;
 import nl.meron.yaeger.engine.entities.entity.Location;
 import nl.meron.yaeger.engine.entities.entity.sprite.SpriteEntity;
 
 public class Logo extends SpriteEntity {
-    public Logo(String resource, Location location, Size size) {
+    public Logo(final String resource, Location location, Size size) {
         super(resource, location, size);
     }
 }
