+++ src/test/java/nl/meron/yaeger/engine/scenes/splash/SplashScreenFactoryTest.java
@@ -1,0 +1,22 @@
+package nl.meron.yaeger.engine.scenes.splash;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+class SplashScreenFactoryTest {
+
+    @Test
+    void callingFactoryCreatesAnInstanceOfSplashScene() {
+        // Arrange
+        var sut = new SplashScreenFactory();
+        var runnable = mock(Runnable.class);
+
+        // Act
+        var yaegerScene = sut.create(runnable);
+
+        // Assert
+        assertTrue(yaegerScene instanceof SplashScene);
+    }
+}
